Move FixedWindowCallback to dd-sdk-android-internal module
Move FixedWindowCallback from dd-sdk-android-rum to dd-sdk-android-internal
(package com.datadog.android.internal.utils) so it can be reused by
session-replay without depending on the whole RUM module.
