+++ features/dd-sdk-android-session-replay/src/main/kotlin/com/datadog/android/sessionreplay/internal/recorder/callback/RecorderWindowCallback.kt
@@ -82,7 +82,13 @@ internal class RecorderWindowCallback(
             )
         }
 
-        return super.dispatchTouchEvent(event)
+        @Suppress("SwallowedException")
+        return try {
+            super.dispatchTouchEvent(event)
+        } catch (e: NullPointerException) {
+            logOrRethrowWrappedCallbackException(e)
+            EVENT_CONSUMED
+        }
     }
 
     override fun onWindowFocusChanged(hasFocus: Boolean) {
@@ -175,9 +181,30 @@ internal class RecorderWindowCallback(
         lastPerformedFlushTimeInNs = timeProvider.getDeviceElapsedTimeNanos()
     }
 
+    private fun logOrRethrowWrappedCallbackException(e: NullPointerException) {
+        // When calling delegate callback, we may have something like
+        // java.lang.NullPointerException: Parameter specified as non-null is null:
+        // method xxx, parameter xxx
+        // This happens because Kotlin delegate expects non-null value incorrectly inferring
+        // non-null type from Java interface definition (seems to be solved in Kotlin 1.8 though)
+        if (e.message?.contains("Parameter specified as non-null is null") == true) {
+            internalLogger.log(
+                InternalLogger.Level.ERROR,
+                InternalLogger.Target.MAINTAINER,
+                { FAIL_TO_PROCESS_MOTION_EVENT_ERROR_MESSAGE },
+                e
+            )
+        } else {
+            @Suppress("ThrowingInternalException") // we need to let client exception to propagate
+            throw e
+        }
+    }
+
     // endregion
 
     companion object {
+        private const val EVENT_CONSUMED: Boolean = true
+
         // every frame we collect the move event positions
         internal val MOTION_UPDATE_DELAY_THRESHOLD_NS: Long =
             TimeUnit.MILLISECONDS.toNanos(16)
@@ -186,5 +213,7 @@ internal class RecorderWindowCallback(
         internal val FLUSH_BUFFER_THRESHOLD_NS: Long = MOTION_UPDATE_DELAY_THRESHOLD_NS * 10
         internal const val MOTION_EVENT_WAS_NULL_ERROR_MESSAGE =
             "RecorderWindowCallback: intercepted null motion event"
+        internal const val FAIL_TO_PROCESS_MOTION_EVENT_ERROR_MESSAGE =
+            "RecorderWindowCallback: wrapped callback failed to handle the motion event"
     }
 }
